authservice/requirements.txt
amqp==5.1.1
asgiref==3.7.1
billiard==4.1.0
celery==5.3.0
cffi==1.15.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.1
Django==4.2.1
django-cors-headers==4.0.0
djangorestframework==3.14.0
kombu==5.3.1
pika==1.3.2
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.7.0
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.6
